Letters moved to app.tsx
